requirements.txt
praw==7.7.1
python-dotenv==1.0.0